# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Base FastAPI",
    "description": "Version 1 of the FastAPI Application",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/app/v1" }],
  "paths": {
...
    "/users/payment-intents": {
      "post": {
        "tags": ["Payment Intents"],
        "summary": "Payment Intents:Create Payment Intent",
        "operationId": "payment_intents_create_payment_intent_users_payment_intents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentIntentCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/users/payment-intents",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/users/payment-intents",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }]
      }
    },
...
    "/organizations/{organization_id}/users": {
      "get": {
        "tags": ["Organization User"],
        "summary": "Organizations:Users",
        "operationId": "organizations_users_organizations__organization_id__users_get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sorted_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "default": "id",
              "title": "Sorted By"
            }
          },
          {
            "name": "sorted_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sorted Order"
            }
          },
          {
            "name": "email__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Email  Icontains"
            }
          },
          {
            "name": "created_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Gte"
            }
          },
          {
            "name": "created_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Lte"
            }
          },
          {
            "name": "updated_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Gte"
            }
          },
          {
            "name": "updated_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Lte"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/users/{user_id}": {
      "get": {
        "tags": ["Organization User"],
        "summary": "Organizations:User",
        "operationId": "organizations_user_organizations__organization_id__users__user_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Id" }
          },
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users/{user_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users/{user_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Organization User"],
        "summary": "Organizations:User:Delete",
        "operationId": "organizations_user_delete_organizations__organization_id__users__user_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Id" }
          },
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users/{user_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users/{user_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/organizations/{organization_id}/payment-intents": {
      "post": {
        "tags": ["Payment Intents"],
        "summary": "Organization::Payment Intents:Create Payment Intent",
        "operationId": "organization__payment_intents_create_payment_intent_organizations__organization_id__payment_intents_post",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentIntentCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/payment-intents",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/payment-intents",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "status": { "type": "string", "title": "Status" },
          "code": { "type": "string", "title": "Code" },
          "title": { "type": "string", "title": "Title" },
          "detail": { "type": "string", "title": "Detail" },
          "source": {
            "anyOf": [
              { "$ref": "#/components/schemas/ErrorSource" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": ["status", "code", "title", "detail"],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": { "type": "string", "title": "Title" },
          "status": { "type": "integer", "title": "Status" },
          "instance": { "type": "string", "title": "Instance" },
          "errors": {
            "items": { "$ref": "#/components/schemas/ErrorDetail" },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": ["title", "status", "instance", "errors"],
        "title": "ErrorResponse"
      },
      "ErrorSource": {
        "properties": {
          "pointer": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Pointer"
          }
        },
        "type": "object",
        "title": "ErrorSource"
      },
      "ListResponseMeta": {
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum number of items to retrieve",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Starting position for retrieval",
            "default": 0
          },
          "sorted_by": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Sorted By",
            "description": "Field name to sort by",
            "default": "id"
          },
          "sorted_order": {
            "type": "string",
            "title": "Sorted Order",
            "description": "Sort order: asc or desc",
            "default": "asc"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items matching the search criteria.",
            "example": 100
          }
        },
        "type": "object",
        "required": ["total_count"],
        "title": "ListResponseMeta"
      },
...
      "PaymentIntentCreate": {
        "properties": {
          "amount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Amount",
            "description": "The amount to be charged in the smallest currency unit (USD).",
            "example": 1000
          }
        },
        "type": "object",
        "required": ["amount"],
        "title": "PaymentIntentCreate"
      },
      "PaymentIntentCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The unique identifier for the payment intent.",
            "example": "pi_1F8Y2e2eZvKYlo2C0d3f4g5h6"
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "description": "The amount to be charged in the smallest currency unit (e.g., cents for USD).",
            "example": "10"
          },
          "amount_inclusive_tax": {
            "type": "string",
            "title": "Amount Inclusive Tax",
            "description": "The amount to be charged inclusive of tax in the smallest currency unit.",
            "example": "11"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "The currency in which the payment intent is created.",
            "example": "usd"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "description": "The client secret used to confirm the payment intent on the client side.",
            "example": "pi_1F8Y2e2eZvKYlo2C0d3f4g5h6_secret_1234567890"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "The current status of the payment intent.",
            "example": "requires_confirmation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "amount_inclusive_tax",
          "currency",
          "client_secret",
          "status"
        ],
        "title": "PaymentIntentCreateResponse"
      },
...
      "UserListRead": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/ListResponseMeta" },
          "data": {
            "items": { "$ref": "#/components/schemas/UserRead" },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["meta", "data"],
        "title": "UserListRead"
      },
      "UserRead": {
        "properties": {
          "id": { "title": "Id" },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "email"],
        "title": "UserRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "auth/jwt/login" } }
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}